requirements.txt
selenium==4.15.0
webdriver-manager==4.0.1
python-dotenv==1.0.0
pydantic==2.5.0
typing-extensions==4.9.0
pyyaml==6.0.1